Subscriptions: dialog re-renderer (#14076)
* dialog re-renderer

* fixing comments

* combining renderers

* adding more attr to sanitizer test
