+++ examples/amp-subscriptions.amp.html
@@ -334,7 +334,8 @@
     Dialog for logged in users
   </template>
   <template type="amp-mustache" subscriptions-dialog subscriptions-display="NOT loggedIn">
-    Dialog for non logged in users
+    <div>Dialog for non logged in users<div>
+    <div subscriptions-display="NOT subscribed">btw u are not subscribed</div>
   </template>
 </body>
 </html>
